+++ packages/app/src/components/PermissionRequest/index.tsx
@@ -12,7 +12,7 @@ export default function PermissionRequest(): React.ReactElement {
   const { permission, requestPermission, subscribe } = useServiceWorkerRegistration();
 
   if (definition.notifications && definition.notifications === 'startup') {
-    if (window.Notification.permission === 'denied') {
+    if (window.Notification?.permission === 'denied') {
       return null;
     }