+++ packages/server/src/api/components/index.js
@@ -1,8 +1,0 @@
-import examples from './examples';
-import parameters from './parameters';
-import requestBodies from './requestBodies';
-import responses from './responses';
-import schemas from './schemas';
-import securitySchemes from './securitySchemes';
-
-export default { examples, parameters, requestBodies, responses, schemas, securitySchemes };