+++ packages/server/src/routes/appRouter/index.js
@@ -1,6 +1,7 @@
-import { partialNormalized } from '@appsemble/utils';
+import { partialNormalized, partialSemver } from '@appsemble/utils';
 
 import tinyRouter from '../../middleware/tinyRouter';
+import blockAssetHandler from './blockAssetHandler';
 import iconHandler from './iconHandler';
 import indexHandler from './indexHandler';
 import manifestHandler from './manifestHandler';
@@ -13,6 +14,12 @@ export default tinyRouter([
   {
     route: /^\/icon-(?<width>\d+)(x(?<height>\d+))?\.(?<format>png|jpg|tiff|webp)$/,
     get: iconHandler,
+  },
+  {
+    route: new RegExp(
+      `^/api/blocks/(?<name>@${partialNormalized.source}/${partialNormalized.source})/versions/(?<version>${partialSemver.source})/(?<filename>.+)$`,
+    ),
+    get: blockAssetHandler,
   },
   {
     route: new RegExp(`^(/${partialNormalized.source})?`),